secretarytools/requirements.txt
Flask==2.3.3
python-docx==0.8.11
flask-cors==4.0.0